Add files via upload
# Java Programs Repository ☕💻

This repository contains all the **Java programs** I write as part of my Java class.  
It starts with the classic *Hello World* and will grow to include more advanced concepts like loops, functions, classes, inheritance, data structures, and more.  

---

## 📂 Repository Structure
Each program is saved as a `.java` file with a name related to the topic.  
For example: